+++ app/src/main/java/com/pepetech/squadhouse/activities/Explore/ExploreActivity.java
@@ -1,4 +1,4 @@
-package com.pepetech.squadhouse.activities;
+package com.pepetech.squadhouse.activities.Explore;
 
 import android.os.Bundle;
 import android.util.Log;
@@ -23,9 +23,9 @@ import com.parse.ParseObject;
 import com.parse.ParseQuery;
 import com.parse.ParseUser;
 import com.pepetech.squadhouse.R;
-import com.pepetech.squadhouse.adapters.ExploreClubAdapter;
-import com.pepetech.squadhouse.adapters.ExploreInterestAdapter;
-import com.pepetech.squadhouse.adapters.ExploreUserAdapter;
+import com.pepetech.squadhouse.activities.Explore.adapters.ExploreClubAdapter;
+import com.pepetech.squadhouse.activities.Explore.adapters.ExploreInterestAdapter;
+import com.pepetech.squadhouse.activities.Explore.adapters.ExploreUserAdapter;
 import com.pepetech.squadhouse.models.Club;
 import com.pepetech.squadhouse.models.Follow;
 import com.pepetech.squadhouse.models.Interest;
