+++ app/src/main/java/com/pepetech/squadhouse/activities/Login/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.pepetech.squadhouse.activities;
+package com.pepetech.squadhouse.activities.Login;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -14,6 +14,7 @@ import com.parse.LogInCallback;
 import com.parse.ParseException;
 import com.parse.ParseUser;
 import com.pepetech.squadhouse.R;
+import com.pepetech.squadhouse.activities.Home.HomeActivity;
 
 public class LoginActivity extends AppCompatActivity {
     public static final String TAG = "LoginActivity";
