+++ app/src/main/java/com/pepetech/squadhouse/activities/Login/WelcomeActivity.java
@@ -1,4 +1,4 @@
-package com.pepetech.squadhouse.activities;
+package com.pepetech.squadhouse.activities.Login;
 
 import android.os.Bundle;
