+++ app/src/main/java/com/pepetech/squadhouse/activities/MyProfile/UpdateUsernameActivity.java
@@ -1,4 +1,4 @@
-package com.pepetech.squadhouse.activities;
+package com.pepetech.squadhouse.activities.MyProfile;
 
 import android.content.Intent;
 import android.os.Bundle;
