+++ app/src/main/res/layout/activity_followers.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:background="@color/eggshell"
     android:layout_height="match_parent"
-    tools:context=".activities.FollowersActivity">
+    tools:context=".activities.Explore.FollowersActivity">
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvFollowers"
         android:layout_width="match_parent"
