+++ app/src/main/res/layout/activity_login.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:layout_margin="16dp"
     android:orientation="vertical"
-    tools:context=".activities.LoginActivity">
+    tools:context=".activities.Login.LoginActivity">
 
     <EditText
         android:id="@+id/etUserName"
