+++ app/src/main/res/layout/activity_settings.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".activities.SettingsActivity">
+    tools:context=".activities.MyProfile.SettingsActivity">
 
     <Button
         android:id="@+id/btnAccount"
