+++ app/src/main/res/layout/activity_sign_up.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".activities.SignUpActivity">
+    tools:context=".activities.Login.SignUpActivity">
 
     <EditText
         android:id="@+id/etUserName"
